+++ Quiz-Website/src/main/java/com/freeuni/proj_100/quizwebsite/model/User.java
@@ -23,7 +23,7 @@ public class User implements UserDetails {
      */
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     /**
      * The unique username used for authentication. Maximum length of 50 characters.
@@ -72,7 +72,7 @@ public class User implements UserDetails {
      * @param createdAt    The creation timestamp
      * @param isAdmin      The administrative status flag
      */
-    public User(Long id, String username, String email, String passwordHash,
+    public User(Integer id, String username, String email, String passwordHash,
                 LocalDateTime createdAt, boolean isAdmin) {
         this.id = id;
         this.username = username;
@@ -100,13 +100,13 @@ public class User implements UserDetails {
         this.createdAt = LocalDateTime.now();
     }
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
     public String getEmail() { return email; }
     public boolean isAdmin() { return isAdmin; }    // JPA expects 'isAdmin' and not 'getIsAdmin'
     public int getTokenVersion() { return tokenVersion; }
     public LocalDateTime getCreatedAt() { return createdAt; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(Integer id) { this.id = id; }
     public void setUsername(String username) { this.username = username; }
     public void setEmail(String email) { this.email = email; }
     public void setPasswordHash(String passwordHash) { this.passwordHash = passwordHash; }
